requirements.txt
# ShariahFolio - Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Deep Learning
torch>=2.0.0

# LangChain & LangGraph
langgraph>=0.0.30
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Portfolio Optimization
scipy>=1.11.0
